refactor(web): align formattable device form with form module conventions
Adapt the form to the patterns established by the storage form refactors
merged from the logical volume form branch:

- Move the validators and the exported validate function from fields.ts
  to a dedicated validations.ts, leaving fields.ts as the field
  vocabulary module (types, constants, defaults).
- Extend the fields.ts re-export to cover the shared field-group types,
  so form-local code imports its whole vocabulary from one module.
- Import FILESYSTEM_ACTION in transformations.ts from fields.ts, since
  shared/transformations.ts no longer re-exports the shared constants.

The fields.test.ts file becomes validations.test.ts: it only exercised
validate, and unlike the partition and logical volume forms there are no
field helpers left to keep a minimal fields.test.ts for.
